+++ src/niprefs/parser/escaper.nim
@@ -1,4 +1,5 @@
-import std/[tables, strutils, parseutils, unicode]
+import std/[strutils, parseutils, unicode, tables]
+import lexer
 
 # https://nim-lang.org/docs/manual.html#lexical-analysis-string-literals
 const escapedTable = {
@@ -38,8 +39,7 @@ proc parseEscapedChar*(str: string, start: Natural = 0): char =
   ##
   ## Check the [manual](https://nim-lang.org/docs/manual.html#lexical-analysis-character-literals) for valid character literals.
   runnableExamples:
-    let text = r"\x23"
-    assert text.parseEscapedChar() == '\x23'
+    assert '\x23' == r"\x23".parseEscapedChar()
 
   if str.len == 0:
     return
@@ -57,12 +57,12 @@ proc parseEscapedChar*(str: string, start: Natural = 0): char =
     elif nextIs('x'):
       var hex: int
       if str.parseHex(hex, 2) != 2:
-        raise newException(Exception, r"Exactly 2 hex decimals are allowed after \x")
+        raise newException(SyntaxError, r"Exactly 2 hex decimals are allowed after \x")
 
       result = char(hex)
 
     else:
-      raise newException(Exception, str & " invalid character constant")
+      raise newException(SyntaxError, str & " invalid character constant")
 
   else:
     result = str[start]
@@ -72,8 +72,8 @@ proc parseEscaped*(str: string): string =
   ##
   ## Check the [manual](https://nim-lang.org/docs/manual.html#lexical-analysis-string-literals) for valid string literals.
   runnableExamples:
-    let text = r"\u1235"
-    assert text.parseEscaped() == "\u1235"
+    assert "\u1235" == r"\u1235".parseEscaped()
+    assert "\x00fd\x0asdsd" == r"\x00fd\x0asdsd".parseEscaped()
 
   var pos = 0
 
@@ -93,8 +93,8 @@ proc parseEscaped*(str: string): string =
       elif nextIs('x'):
         var hex: int
 
-        if str.parseHex(hex, pos+2) != 2:
-          raise newException(Exception, r"Exactly 2 hex decimals are allowed after \x")
+        if str.parseHex(hex, pos+2, maxLen = 2) != 2:
+          raise newException(SyntaxError, r"Exactly 2 hex decimals are allowed after \x")
 
         pos += 3 # 2 + 1
         result &= char(hex)
@@ -103,21 +103,23 @@ proc parseEscaped*(str: string): string =
         var hex: int
         pos += str.parseHex(hex, pos+3) + 3
         if pos > str.high or str[pos] != '}':
-          raise newException(Exception, r"Missing closing } for \u{H+}")
+          raise newException(SyntaxError, r"Missing closing } for \u{H+}")
 
+        if hex > 0x10FFFF:
+          raise newException(SyntaxError, "Unicode codepoint must be lower than 0x10FFFF, but was: " & $hex)
         result &= Rune(hex)
 
       elif nextIs('u'):
         var hex: int
 
-        if str.parseHex(hex, pos+2) != 4:
-          raise newException(Exception, r"Exactly 4 hex decimals are allowed after \u")
+        if str.parseHex(hex, pos+2, maxLen = 4) != 4:
+          raise newException(SyntaxError, r"Exactly 4 hex decimals are allowed after \u")
 
         pos += 5 # 4 + 1
         result &= Rune(hex)
 
       else:
-        raise newException(Exception, str[pos..pos+1] & " invalid character constant")
+        raise newException(SyntaxError, str[pos..pos+1] & " invalid character constant")
 
     else:
       result &= chr